+++ apps/trackers/exceptions.py
@@ -11,6 +11,12 @@ class NoPriorityAvailableError(BTSException):
     """exception class for missing correct priority corresponding to Impact"""
 
 
+class TrackerCreationError(BTSException):
+    """
+    exception class for the cases of unsuccessful tracker creation
+    """
+
+
 class UnsupportedTrackerError(BTSException):
     """
     exception class for the cases of the unsupported tracker filing